requirements.txt
annotated-types==0.7.0
anyio==3.7.1
async-timeout==5.0.1
asyncpg==0.29.0
bcrypt==5.0.0
blis==0.7.11
catalogue==2.0.10
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.16.0
confection==0.1.2
cryptography==46.0.3
cymem==2.0.8
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.104.1
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==3.0.3
murmurhash==1.0.10
numpy==1.24.4
packaging==25.0
passlib==1.7.4
pathy==0.10.3
Pillow==10.1.0
preshed==3.0.9
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.23
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
PyJWT==2.10.1
pytesseract==0.3.13
python-dotenv==1.0.0
python-http-client==3.3.7
python-jose==3.3.0
PyYAML==6.0.3
requests==2.31.0
rsa==4.9.1
sendgrid==6.12.0
six==1.17.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.23
srsly==2.4.8
starkbank-ecdsa==2.2.0
starlette==0.27.0
thinc==8.1.12
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.24.0
uvloop==0.22.1
wasabi==1.1.3
watchfiles==1.1.1
weasel==0.3.4
websockets==15.0.1
Werkzeug==3.1.4
